Businesses Model Wrapper added and Root Store revamped
