+++ src/models/Business/index.ts
@@ -6,5 +6,4 @@ export * from './LocationModel';
 export * from './MessagingModel';
 export * from './OpenHoursModel';
 export * from './PhotoDetailsModel';
-export * from './ErrorModel';
 export * from './OpenHoursModel';
